Check for file existence before attempting to open it

+++ src/action/base/create_file.rs
@@ -266,6 +266,10 @@ impl Action for CreateFile {
             buf: _,
             force: _,
         } = self;
+        // The user already deleted it
+        if !path.exists() {
+            return Ok(())
+        }
 
         remove_file(&path)
             .await
